+++ tars/movement/__init__.py
@@ -1,2 +1,4 @@
 from .driver import ServoDriver
 from .gaits import Gaits
+
+__all__ = ["ServoDriver", "Gaits"]